Remade Nikon drivers

+++ CameraControl/Devices/ICameraDevice.cs
@@ -9,7 +9,7 @@ namespace CameraControl.Devices
     PropertyValue<int> FNumber { get; set; }
     PropertyValue<int> IsoNumber { get; set; }
     PropertyValue<long> ShutterSpeed { get; set; }
-    PropertyValue<int> WhiteBalance { get; set; }
+    PropertyValue<long> WhiteBalance { get; set; }
     PropertyValue<uint> Mode { get; set; }
     PropertyValue<int> ExposureCompensation { get; set; }
     PropertyValue<int> CompressionSetting { get; set; }
